refactor(config): move openai functions to settings mixin
see #44
