+++ odoogpt/models/res_config_settings.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from odoo import models, fields, api, _
-from odoo.exceptions import UserError
 
 
 class ResConfigSettings(models.TransientModel):
@@ -48,29 +47,3 @@ class ResConfigSettings(models.TransientModel):
             odoogpt_openai_prompt_suffix = self.env.company.odoogpt_openai_prompt_suffix
         )
         return res
-
-
-    def odoogpt_openai_test(self):
-        """Call Models list api to check everything is properly set up"""
-        OdoogptOpenaiUtils = self.env['odoogpt.openai.utils']
-        try:
-            OdoogptOpenaiUtils.models_list()
-        except Exception as ex:
-            raise UserError(ex)
-
-        return {
-            'type': 'ir.actions.client',
-            'tag': 'display_notification',
-            'params': {
-                'title': _('Test success!'),
-                'message': _('Everything properly set up! You\'re good to go!'),
-                'sticky': False,
-            }
-        }
-
-
-    def odoogpt_openai_model_select_from_db(self):
-        """Get Models from OpenAI api and show selector wizart"""
-        self.env['odoogpt.openai.model'].sudo().refresh_from_api(format='model')
-        return self.env.ref('odoogpt.odoogpt_openai_model_select_wizard_act_window').read()[0]
-
