+++ odoogpt/views/res_config_settings.xml
@@ -8,12 +8,7 @@
             <field name="inherit_id" ref="base.res_config_settings_view_form" />
             <field name="arch" type="xml">
                 <xpath expr="//div[hasclass('settings')]" position="inside">
-                    <div
-                        class="app_settings_block"
-                        data-string="Odoo GPT"
-                        string="Odoo GPT"
-                        data-key="odoogpt"
-                    >
+                    <div class="app_settings_block" data-string="Odoo GPT" string="Odoo GPT" data-key="odoogpt">
                         <div id="odoogpt_openai_api">
                             <h2>OpenAI Api</h2>
 
@@ -79,7 +74,7 @@
 
                         <div id="odoogpt_chat_customization">
                             <h2>OdooGPT Chat Customization</h2>
-    
+
                             <div class="row mt16 o_settings_container" name="odoogpt_chat_customization_container">
                                 <div class="col-12 o_setting_box" id="odoogpt_chat_method">
                                     <div class="o_setting_left_pane">
@@ -150,7 +145,6 @@
                                         <field name="odoogpt_chat_system_message" required="1" />
                                     </div>
                                 </div>
-
                             </div>
                         </div>
                     </div>
